core_api/requirements.txt
Django==5.0.6
django-environ
psycopg==3.1.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2